organize mocha tests
