+++ app/scripts/mocha-tests/account-tracker.test.js
@@ -4,7 +4,7 @@ import EventEmitter from 'events';
 import { SINGLE_CALL_BALANCES_ADDRESS } from '../constants/contracts';
 
 import { createTestProviderTools } from '../../../test/stub/provider';
-import AccountTracker from './account-tracker';
+import AccountTracker from '../lib/account-tracker';
 
 const noop = () => true;
 const currentNetworkId = '5';
