+++ app/scripts/mocha-tests/ens.test.js
@@ -1,6 +1,6 @@
 import { strict as assert } from 'assert';
 import sinon from 'sinon';
-import EnsController from '.';
+import EnsController from '../controllers/ens';
 
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 const ZERO_X_ERROR_ADDRESS = '0x';
